fix: modified deleteNeed method http status

+++ src/main/java/com/soprasteria/booking/controller/impl/NeedControllerImpl.java
@@ -62,8 +62,14 @@ public class NeedControllerImpl implements NeedController {
 
     @Override
     public ResponseEntity<Void> deleteNeed(Long id) {
-        needService.deleteNeed(id);
-        return new ResponseEntity(HttpStatus.OK);
+        NeedDTO needDTO = needService.findById(id);
+        if(needDTO==null){
+            return ResponseEntity.noContent().build();
+        }
+        else{
+            needService.deleteNeed(id);
+            return new ResponseEntity(HttpStatus.OK);
+        }
     }
